+++ python-packages/contract_wrappers/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.0.0 - TBD
+
+-   Updated for version 3 of the protocol.
+-   Allow wrappers to be instantiated with EITHER a Web3.py `BaseProvider` OR an already-instantiated `Web3` client object.
+-   Accept `str`ing arguments to `bytes` contract method parameters.
+-   Expanded documentation examples.
+-   Moved methods `jsdict_to_order()` and `order_to_jsdict()` from `zero_ex.contract_wrappers.exchange.types` to `zero_ex.contract_wrappers.order_conversions`.
+-   Changed field name `zero_ex.contract_wrappers.tx_params.TxParams.gasPrice` to `.gas_price`.
+
 ## 1.1.0 - 2019-08-14
 
 -   Added wrapper for DevUtils contract.